Move static functions in to utils.
