+++ src/services/role/utils/filterByQuery.js
@@ -1,0 +1,16 @@
+const sift = require('sift');
+
+const filterByQuery = (roles = [], query = {}) => {
+	let result = roles;
+	const q = { ...query };
+	delete q.$skip;
+	delete q.$limit;
+	Object.entries(q).forEach(([key, v]) => {
+		if (v instanceof RegExp) {
+			result = result.filter((role) => role[key].match(v));
+		}
+	});
+	return result.filter(sift(q));
+};
+
+module.exports = filterByQuery;
