+++ src/services/role/utils/index.js
@@ -1,0 +1,17 @@
+const {
+	preparedRoles,
+	dissolveInheritPermission,
+	addDisplayName,
+} = require('./preparedRoles');
+const filterByQuery = require('./filterByQuery');
+const paginate = require('./paginate');
+const unique = require('./unique');
+
+module.exports = {
+	preparedRoles,
+	dissolveInheritPermission,
+	addDisplayName,
+	filterByQuery,
+	paginate,
+	unique,
+};
