+++ src/services/role/utils/unique.js
@@ -1,0 +1,3 @@
+const unique = (...permissions) => ([...new Set(Array.prototype.concat.apply([], permissions))]);
+
+module.exports = unique;
